An exciting memory game.
This is a simple game which goal is for user to repeat the pattern showed by the program on the screen. With each round a new step is added to the pattern making, the game becomes much more difficult with every round. Additionally, the program plays a different sound for each field that is activated. On top Of that if the user selects wrong field during pattern making then the game is over and need to re-start from beginning. The provided example constituted of 4 different fields with different color for each field.